# 修正woc openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "修正woc",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/captain/team": {
      "post": {
        "summary": "更新队伍信息（必做）",
        "deprecated": false,
        "description": "",
        "tags": [
          "captain"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Team"
              },
              "example": ""
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "errCode": {
                      "type": "integer"
                    },
                    "errMsg": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/TeamModel"
                    }
                  },
                  "required": [
                    "success",
                    "data",
                    "errMsg",
                    "errCode"
                  ],
                  "x-apifox-orders": [
                    "success",
                    "errCode",
                    "errMsg",
                    "data"
                  ],
                  "x-apifox-ignore-properties": []
                }
              }
            },
            "headers": {},
            "x-apifox-name": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-apifox-folder": "修正woc/captain",
        "x-apifox-status": "released",
        "x-run-in-apifox": "https://app.apifox.com/web/project/5750446/apis/api-263845415-run"
      },
      "get": {
        "summary": "获取队伍信息（必做）",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "errCode": {
                      "type": "integer"
                    },
                    "errMsg": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/TeamModel"
                    }
                  },
                  "required": [
                    "success",
                    "errCode",
                    "errMsg",
                    "data"
                  ],
                  "x-apifox-orders": [
                    "success",
                    "errCode",
                    "errMsg",
                    "data"
                  ],
                  "x-apifox-ignore-properties": []
                }
              }
            },
            "headers": {},
            "x-apifox-name": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-apifox-folder": "修正woc",
        "x-apifox-status": "released",
        "x-run-in-apifox": "https://app.apifox.com/web/project/5750446/apis/api-263845399-run"
      }
    },
    "/captain/team/{teamId}/member": {
      "post": {
        "summary": "添加成员（必做）",
        "deprecated": false,
        "description": "",
        "tags": [
          "captain"
        ],
        "parameters": [
          {
            "name": "teamId",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Member"
              },
              "example": ""
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-apifox-folder": "修正woc/captain",
        "x-apifox-status": "released",
        "x-run-in-apifox": "https://app.apifox.com/web/project/5750446/apis/api-263845416-run"
      },
      "delete": {
        "summary": "删除成员（必做）",
        "deprecated": false,
        "description": "",
        "tags": [
          "captain"
        ],
        "parameters": [
          {
            "name": "teamId",
            "in": "path",
            "description": "",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer",
                    "description": "member id"
                  }
                },
                "required": [
                  "id"
                ],
                "x-apifox-orders": [
                  "id"
                ],
                "x-apifox-ignore-properties": []
              },
              "example": ""
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "errCode": {
                      "type": "integer"
                    },
                    "errMsg": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {},
                      "x-apifox-orders": [],
                      "x-apifox-ignore-properties": []
                    }
                  },
                  "required": [
                    "success",
                    "errCode",
                    "errMsg",
                    "data"
                  ],
                  "x-apifox-orders": [
                    "success",
                    "errCode",
                    "errMsg",
                    "data"
                  ],
                  "x-apifox-ignore-properties": []
                }
              }
            },
            "headers": {},
            "x-apifox-name": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-apifox-folder": "修正woc/captain",
        "x-apifox-status": "released",
        "x-run-in-apifox": "https://app.apifox.com/web/project/5750446/apis/api-263845417-run"
      },
      "patch": {
        "summary": "修改成员信息（必做）",
        "deprecated": false,
        "description": "",
        "tags": [
          "captain"
        ],
        "parameters": [
          {
            "name": "teamId",
            "in": "path",
            "description": "",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Member"
              },
              "example": ""
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalResponse"
                }
              }
            },
            "headers": {},
            "x-apifox-name": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-apifox-folder": "修正woc/captain",
        "x-apifox-status": "released",
        "x-run-in-apifox": "https://app.apifox.com/web/project/5750446/apis/api-263845418-run"
      },
      "get": {
        "summary": "获取成员列表（必做）",
        "deprecated": false,
        "description": "",
        "tags": [
          "captain"
        ],
        "parameters": [
          {
            "name": "teamId",
            "in": "path",
            "description": "",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "errCode": {
                      "type": "integer"
                    },
                    "errMsg": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "studentId": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "academy": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "studentId",
                          "name",
                          "academy"
                        ],
                        "x-apifox-orders": [
                          "studentId",
                          "name",
                          "academy"
                        ],
                        "x-apifox-ignore-properties": []
                      }
                    }
                  },
                  "required": [
                    "success",
                    "errCode",
                    "errMsg",
                    "data"
                  ],
                  "x-apifox-orders": [
                    "success",
                    "errCode",
                    "errMsg",
                    "data"
                  ],
                  "x-apifox-ignore-properties": []
                }
              }
            },
            "headers": {},
            "x-apifox-name": "成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-apifox-folder": "修正woc/captain",
        "x-apifox-status": "released",
        "x-run-in-apifox": "https://app.apifox.com/web/project/5750446/apis/api-263845419-run"
      }
    },
...
  },
  "components": {
    "schemas": {
      "Team": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "comId": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "captainId": {
            "type": "integer"
          },
          "captainName": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "memberIds": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "instructorIds": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "createTime": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "comId",
          "name",
          "captainId",
          "captainName",
          "memberIds",
          "instructorIds",
          "status",
          "createTime"
        ],
        "x-apifox-orders": [
          "id",
          "comId",
          "name",
          "captainId",
          "captainName",
          "status",
          "memberIds",
          "instructorIds",
          "createTime"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": "Schemas"
      },
      "TeamModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "comId": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "captainId": {
            "type": "integer"
          },
          "captainName": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "memberNames": {
            "type": "string",
            "description": "学号加姓名格式\n例如：B24XXXXXX某某某"
          },
          "instructorNames": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "comId",
          "name",
          "captainId",
          "captainName",
          "memberNames",
          "instructorNames",
          "status"
        ],
        "x-apifox-orders": [
          "id",
          "comId",
          "name",
          "captainId",
          "captainName",
          "status",
          "memberNames",
          "instructorNames"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": "Schemas"
      },
      "GlobalResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "errCode": {
            "type": "integer"
          },
          "errMsg": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {},
            "x-apifox-orders": [],
            "x-apifox-ignore-properties": []
          }
        },
        "required": [
          "success",
          "data",
          "errMsg",
          "errCode"
        ],
        "x-apifox-orders": [
          "success",
          "errCode",
          "errMsg",
          "data"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": "Schemas"
      },
...
      "Member": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "studentId": {
            "type": "string"
          },
          "teamId": {
            "type": "integer"
          },
          "academyId": {
            "type": "integer"
          },
          "phone": {
            "type": "string"
          },
          "isCaptain": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "phone",
          "academyId",
          "teamId",
          "studentId",
          "name",
          "isCaptain"
        ],
        "x-apifox-orders": [
          "id",
          "name",
          "studentId",
          "teamId",
          "academyId",
          "phone",
          "isCaptain"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": "Schemas"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [],
  "security": []
}